Fix eval test
