+++ packages/snaps-utils/src/eval-worker.ts
@@ -3,9 +3,9 @@ import 'ses/lockdown';
 
 import { readFileSync } from 'fs';
 
-import type { HandlerType } from './handlers';
+import type { HandlerType } from './handler-types';
+import { SNAP_EXPORT_NAMES } from './handler-types';
 import { generateMockEndowments } from './mock';
-import { SNAP_EXPORT_NAMES } from './types';
 
 declare let lockdown: any, Compartment: any;
