+++ packages/snaps-utils/src/handler-types.ts
@@ -1,0 +1,35 @@
+export enum HandlerType {
+  OnRpcRequest = 'onRpcRequest',
+  OnTransaction = 'onTransaction',
+  OnCronjob = 'onCronjob',
+  OnInstall = 'onInstall',
+  OnUpdate = 'onUpdate',
+  OnNameLookup = 'onNameLookup',
+  OnKeyringRequest = 'onKeyringRequest',
+}
+
+export type SnapHandler = {
+  /**
+   * The type of handler.
+   */
+  type: HandlerType;
+
+  /**
+   * Whether the handler is required, i.e., whether the request will fail if the
+   * handler is called, but the snap does not export it.
+   *
+   * This is primarily used for the lifecycle handlers, which are optional.
+   */
+  required: boolean;
+
+  /**
+   * Validate the given snap export. This should return a type guard for the
+   * handler type.
+   *
+   * @param snapExport - The export to validate.
+   * @returns Whether the export is valid.
+   */
+  validator: (snapExport: unknown) => boolean;
+};
+
+export const SNAP_EXPORT_NAMES = Object.values(HandlerType);
